Add files via upload
This project consists of 4 parts (part 2, part 3, part 4 are included here as part 1 was the star uml analysis of the project). So in part 2 you can see the implementation of some of the basic classes. Afterwards in part 3 I had to change some classes so they use dynamic memory allocation and pointers and upgrade some of the classes so the game is even more capable. At the last part, part 4 I used I imlemented specific libraries and another project class "Computerplayer" so that the game becomes even more advanced. I will briefly analyse my grade in every part of the project. So I got a 9/10 in the uml analysis, a 10/10 in part 2, a 9/10 in part 3 and a 10/10 in part 4. Also almost forgot, the project is entirely created in C++ using inheritance and other functionalities of the programming language.